Unload form and move the frequency control logic to separate component
